Draft implementation

+++ internal/backend/models/settings/settings.go
@@ -4,7 +4,7 @@ type Settings struct {
 	BaseUrl               string
 	Headers               map[string]string
 	ModelName             string
-	Temperature           float32
+	Temperature           float64
 	DefaultInputLanguage  string
 	DefaultOutputLanguage string
 	Languages             []string
